Refactored refresh controller.
